Move distinct components to individual files
